+++ src/App/Model/Mineral.hs
@@ -3,7 +3,7 @@ module App.Model.Mineral where
 import App.Prelude
 
 data Mineral = Mineral
-  { available :: Double
+  { available :: Int
   , accessibility :: Double
   }
   deriving (Show, Generic)
